Fix - Fixed the int type error when using json dump

+++ src/emhass/retrieve_hass.py
@@ -200,7 +200,7 @@ class retrieve_hass:
         list_df = copy.deepcopy(data_df).loc[data_df.index[idx]:].reset_index()
         list_df.columns = ['timestamps', entity_id]
         ts_list = [str(i) for i in list_df['timestamps'].tolist()]
-        vals_list = [np.round(i) for i in list_df[entity_id].tolist()]
+        vals_list = [int(np.round(i)) for i in list_df[entity_id].tolist()]
         forecast_list = []
         for i, ts in enumerate(ts_list):
             datum = {}
